requirements.txt
langchain
langchain_community
openai
huggingface_hub
streamlit
python-dotenv